sample-project/requirements.txt
anthropic>=0.42.0
psycopg2-binary>=2.9.9
rich>=13.7.0
